pdf-parser-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
camelot-py[cv]==0.11.0
PyMuPDF==1.23.8
pymysql==1.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
opencv-python-headless==4.8.1.78
paddleocr>=2.9.0
paddlepaddle>=2.6.0
